Fix: 상세 페이지 분리
